Move skip utilities
